requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.3.0
bcrypt==4.2.1
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.104.1
filetype==1.2.0
frozenlist==1.7.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.63
langchain-google-genai==2.1.5
langchain-openai==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.147
marshmallow==3.26.1
multidict==6.6.3
mypy-extensions==1.1.0
numpy==1.26.4
openai==1.99.2
orjson==3.11.1
packaging==23.2
passlib==1.7.4
passlib[bcrypt]==1.7.4
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
python-dotenv==1.0.0
python-multipart==0.0.6
pyyaml==6.0.2
regex==2025.7.34
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
sqlalchemy==2.0.42
starlette==0.27.0
tenacity==8.5.0
tiktoken==0.10.0
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.24.0
yarl==1.20.1
zstandard==0.23.0
